Moved Conn interface to its own file
