+++ src/github.com/cjlucas/unnamedcast/koda/adapter.go
@@ -8,31 +8,7 @@ import (
 	"gopkg.in/redis.v3"
 )
 
-type ZRangeByScoreOpts struct {
-	Min          float64
-	Max          float64
-	MinInclusive bool
-	MaxInclusive bool
-	Offset       int64
-	Count        int64
-}
-
 var NilError = redis.Nil
-
-type Conn interface {
-	Incr(key string) (int, error)
-	HIncr(key, field string) (int, error)
-	HGet(key, field string) (string, error)
-	HGetAll(key string) ([]string, error)
-	HSet(key, field, value string) (bool, error)
-	RPush(key string, value ...string) (int, error)
-	LPush(key string, value ...string) (int, error)
-	BRPop(timeout time.Duration, keys ...string) ([]string, error)
-	ZAddNX(key string, score float64, member string) (int, error)
-	ZRem(key string, members ...string) (int, error)
-	ZRangeByScore(key string, opt *ZRangeByScoreOpts) ([]string, error)
-	Close() error
-}
 
 // GoRedisAdapter is an adapter for the redis.v3 library
 type GoRedisAdapter struct {
